Remove rule_sets construct
Restrict the processor to a single array of rules. This is similar to the structure of the signature utilities, but it instantiates the rules with configuration. This moves rules out of configuration, and out of the required_config population change. To keep the rules happy, all necessary configuration for rules has moved to the Processor_2015 class and is passed to rules at initialization in the `config` variable. Required config options are left on the rules to support testing.